make Tx a static class
